producer/requirements.txt
ultralytics
opencv-python
kafka-python
numpy
torch
